Restructure research routes into overview and content groups
