+++ src/app/research/(content)/digital-rights-ai/page.tsx
