+++ src/app/research/(content)/societal-impacts/page.tsx
