+++ src/app/research/(overview)/layout.tsx
@@ -1,0 +1,25 @@
+import type { ReactNode } from 'react';
+import type { Metadata } from 'next';
+
+import AmbientLayoutFrame from '@/components/AmbientLayoutFrame';
+import ResearchHero from '../ResearchHero';
+import { RESEARCH_BACKGROUND_CLASS, RESEARCH_OVERLAY_CLASS } from '../ambientConfig';
+
+export const metadata: Metadata = {
+  title: 'Research — AI Ethics, Agentic Systems, Privacy',
+  description:
+    'Research highlights in AI ethics, agentic systems, and privacy-preserving technologies at Traceremove Research.',
+};
+
+export default function ResearchOverviewLayout({ children }: { children: ReactNode }) {
+  return (
+    <AmbientLayoutFrame
+      variant="research"
+      backgroundClassName={RESEARCH_BACKGROUND_CLASS}
+      overlayClassName={RESEARCH_OVERLAY_CLASS}
+      hero={<ResearchHero />}
+    >
+      {children}
+    </AmbientLayoutFrame>
+  );
+}
