+++ src/app/research/ambientConfig.ts
@@ -1,0 +1,4 @@
+export const RESEARCH_BACKGROUND_CLASS = 'opacity-70';
+
+export const RESEARCH_OVERLAY_CLASS =
+  'bg-[radial-gradient(circle_at_12%_18%,rgba(56,189,248,0.18),transparent_58%),radial-gradient(circle_at_82%_12%,rgba(139,92,246,0.22),transparent_60%),linear-gradient(180deg,rgba(2,6,23,0.96)_0%,rgba(15,23,42,0.93)_48%,rgba(15,23,42,0.95)_100%)]';
